requirements.txt
PyTelegramBotAPI
python-dotenv